removed the old tracking module, and replaced with the new. updated the relating tests also
