+++ observatory/tracking.py
@@ -1,212 +1,20 @@
-"""
-The tracking module is used to record various parts of your machine learning runs.
-You can use the :func:`start_run <observatory.tracking.start_run>` method from 
-this module to start tracking metrics, outputs and settings for your 
-machine learning runs.
-
-If you want to customize tracking behavior you can create your own instance
-of :class:`TrackingClient <observatory.tracking.TrackingClient>` to send 
-information to the tracking API directly. Please note that this means you have 
-to implement session semantics yourself.
-"""
-
+from abc import ABC, abstractmethod
 import json
 import re
 import warnings
 from os import path
 from time import time
 from uuid import uuid4
+import inspect
 
 import requests
-from observatory import settings
+from observatory import settings, benchmark_local_saving
 from observatory.constants import LABEL_PATTERN
 
-
-class TrackingClient:
-    """
-    Implements logic to convert raw data into HTTP requests to the tracking endpoint.
-
-    Typically you should not be using this class in your own code. Instead make use of the `start_run` method
-    in the same module to start a new tracking session. This ensures correct tracking behavior in your application.
-
-    This client is meant as a wrapper around all the HTTP related logic from the tracking session's point of view.
-    """
-
-    def __init__(self, url):
-        """
-        Initializes a new instance of the tracking client
-
-        Parameters:
-        -----------
-        url : str
-            The URL of the tracking endpoint to communicate with
-        """
-        self.url = url
-
-    def record_session_start(self, model, version, experiment, run_id):
-        """
-        Records the start of a session.
-
-        This method sends a HTTP request with the right payload to the observatory tracking endpoint.
-        The result is a 201 when the server succesfully recorded the session completion. Otherwise
-        the server will return a 500 response.
-
-        Parameters:
-        -----------
-        model : str
-            The name of the model
-        version : int
-            The version of the model
-        experiment : str
-            The name of the experiment
-        run_id : str
-            The identifier for the run
-
-        Returns:
-        --------
-        requests.Response
-            The response from the server
-        """
-        handler_url = f'{self.url}/api/models/{model}/versions/{version}/experiments/{experiment}/runs'
-        return requests.post(handler_url, json={'run_id': run_id})
-
-    def record_settings(self, model, version, experiment, run_id, settings):
-        """
-        Records the settings of an experiment run.
-
-        This method sends a HTTP request with the right payload to the observatory tracking endpoint.
-        The result is a 201 when the server succesfully recorded the session completion. Otherwise
-        the server will return a 500 response.
-
-        Parameters:
-        -----------
-        model : str
-            The name of the model
-        version : int
-            The version of the model
-        experiment : str
-            The name of the experiment
-        run_id : str
-            The identifier for the run
-        settings : dict
-            The dictionary with run settings
-
-        Returns:
-        --------
-        requests.Response
-            The response from the server
-        """
-        handler_url = f'{self.url}/api/models/{model}/versions/{version}/experiments/{experiment}/runs/{run_id}/settings'
-        return requests.post(handler_url, json=settings)
-
-    def record_metric(self, model, version, experiment, run_id, metric_name, metric_value):
-        """
-        Records a metric during a run.
-
-        This method sends a HTTP request with the right payload to the observatory tracking endpoint.
-        The result is a 201 when the server succesfully recorded the session completion. Otherwise
-        the server will return a 500 response.
-
-        Parameters:
-        -----------
-        model : str
-            The name of the model
-        version : int
-            The version of the model
-        experiment : str
-            The name of the experiment
-        run_id : str
-            The identifier for the run
-        metric_name : str
-            The name of the metric
-        metric_value : str
-            The value of the metric
-
-        Returns:
-        --------
-        requests.Response
-            The response from the server
-        """
-        handler_url = f'{self.url}/api/models/{model}/versions/{version}/experiments/{experiment}/runs/{run_id}/metrics'
-        return requests.post(handler_url, json={'name': metric_name, 'value': metric_value})
-
-    def record_output(self, model, version, experiment, run_id, filename, file):
-        """
-        Records an output of an experiment run
-
-        This method sends a HTTP request with the right payload to the observatory tracking endpoint.
-        The result is a 201 when the server succesfully recorded the session completion. Otherwise
-        the server will return a 500 response.
-
-        Parameters:
-        -----------
-        model : str
-            The name of the model
-        version : int
-            The version of the model
-        experiment : str
-            The name of the experiment
-        run_id : str
-            The identifier for the run
-        filename : str
-            The filename of the output
-        file : object
-            The file handle to use for reading the output data
-
-        Returns:
-        --------
-        requests.Response
-            The response from the server
-        """
-        handler_url = f'{self.url}/api/models/{model}/versions/{version}/experiments/{experiment}/runs/{run_id}/outputs/{filename}'
-
-        file_collection = {
-            'file': (filename, file, 'application/octet-stream')
-        }
-
-        return requests.put(handler_url, files=file_collection)
-
-    def record_session_end(self, model, version, experiment, run_id, status):
-        """
-        Records the end of a session.
-
-        This method sends a HTTP request with the right payload to the observatory tracking endpoint.
-        The result is a 201 when the server succesfully recorded the session completion. Otherwise
-        the server will return a 500 response.
-
-        Parameters:
-        -----------
-        model : str
-            The name of the model
-        version : int
-            The version of the model
-        experiment : str
-            The name of the experiment
-        run_id : str
-            The identifier for the run
-        status : str
-            The status of the run
-
-        Returns:
-        --------
-        requests.Response
-            The response from the server
-        """
-        handler_url = f'{self.url}/api/models/{model}/versions/{version}/experiments/{experiment}/runs/{run_id}'
-        return requests.put(handler_url, json={'status': status})
-
-
 class TrackingSession:
-    """
-    The tracking session is used to track data related to a single training run.
+    #trackingsession
 
-    When you invoke :func:`start_run <observatory.start_run>` this class is instantiated for you
-    with the correct settings to start tracking data.
-
-    Any server connection used by the tracking logic is automatically opened and closed for you.
-    """
-
-    def __init__(self, name, version, experiment, run_id, tracking_client):
+    def __init__(self, name, version, experiment, run_id):
         """
         Initializes the tracking session with the necessary tracking information
         and a pre-initialized tracking client for recording the actual metrics.
@@ -221,14 +29,186 @@ class TrackingSession:
             Name of the experiment
         run_id : string
             ID of the run
-        tracking_client : object
-            Instance of the tracking service client
         """
         self.name = name
         self.version = version
-        self.run_id = run_id
-        self.tracking_client = tracking_client
         self.experiment = experiment
+        self.run_id = run_id
+        self._state = LocalState()
+    
+    def change(self, state):
+        self._state.switch(state)
+
+    def record_metric(self, name, value):
+        """
+        Records a metric value on the server
+
+        Parameters
+        ----------
+        name : string
+            The name of the metric to record
+        value : float
+            The value of the metric to records
+        """
+
+        # Typechecking in python is a no-go under normal circumstances.
+        # But here we're using it, because the server expects a string and float.
+
+        if name is None or type(name) != str or name.strip() == '':
+            raise AssertionError('Please provide a valid name for the metric.')
+
+        if not re.match(LABEL_PATTERN, name):
+            raise AssertionError(
+                'Please provide a valid name for the metric.' +
+                'it can contain lower-case alpha-numeric characters and dashes only.')
+
+        if value is None or (type(value) != float and type(value) != int):
+            raise AssertionError(
+                'Please provide a valid value for the metric.')
+
+        self._state.record_metric(
+            self.name, self.version, self.experiment, self.run_id, name, value)
+
+    def record_settings(self, **settings):
+        """
+        Records settings used for the run
+
+        Parameters
+        ----------
+        settings : object
+            A dictionary containing all settings used for the run.
+            This can be passed in as `key=value` pairs.
+        """
+        
+        if settings is None:
+            warnings.warn(
+                'Trying to record empty settings. ' + 
+                'To prevent risk of settings loss on the server, ' +
+                'the empty settings collection is discarded.', 
+                RuntimeWarning)
+
+        self._state.record_settings(
+            self.name, self.version, self.experiment,
+            self.run_id, dict(settings))
+
+    def record_output(self, input_file, filename):
+        """
+        Records an output for the current run.
+
+        Parameters
+        ----------
+        input_file : object
+            Filename or handle to input file
+        filename : string
+            Name of the file as it should be stored on the server
+        """
+
+        if filename is None or filename.strip() == '':
+            raise AssertionError(
+                'Please provide a valid filename to store the output on the server.')
+
+        if input_file is None or input_file.strip() == '':
+            raise AssertionError(
+                'Please provide a valid filename for input_file.')
+
+        absolute_file_path = path.abspath(input_file)
+
+        if not path.exists(absolute_file_path):
+            raise AssertionError(
+                f'Could not find source file {absolute_file_path} ' +
+                'to upload as output of this run. Please make ' +
+                'sure that the file exists on disk.')
+
+        self._state.record_output(
+            self.name, self.version, self.experiment,
+            self.run_id, filename, open(absolute_file_path))
+
+    def __enter__(self):
+        self._state.record_session_start(
+            self.name, self.version, self.experiment, self.run_id)
+
+        return self
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        if exc_type is None:
+            session_status = 'COMPLETED'
+        else:
+            session_status = 'FAILED'
+
+        self._state.record_session_end(
+            self.name, self.version, self.experiment,
+            self.run_id, session_status)
+
+
+        return exc_type is None
+
+class ObservatoryState(ABC):
+    def __init__(self):
+        self.n = None
+
+    def switch(self, state):
+        self.__class__ = state
+
+    @abstractmethod
+    def record_metric(self, name, value):
+        """
+        This method will record a metric.
+        """
+        pass
+
+    @abstractmethod
+    def record_settings(self, name, value):
+        """
+        This method will record a setting.
+        """
+        pass
+
+    @abstractmethod
+    def record_output(self, name, value):
+        """
+        This method will record a output.
+        """
+        pass
+
+    @abstractmethod
+    def record_session_start(self, model, version, experiment, run_id):
+        """
+        This method will record the start of a run.
+        """
+        pass
+
+    @abstractmethod
+    def record_session_end(self, model, version, experiment, run_id, status):
+        """
+        This method will record the end of a run.
+        """
+        pass
+
+class LocalState(ObservatoryState):
+    #Handels metrics for the local filesystem -> Sink.py
+
+    def record_metric(self, model, version, experiment, run_id, name, value):
+        #sink.save_metric(model, version, experiment, run_id, name, value)
+        #localstate is nothing more than a nice data handler that passes is to sink.py
+        #this is because the sever is also going to use sink.py to save data
+        print("LocalState : record_metric")
+        benchmark_local_saving.benchmark_text_record_metric(self, model, version, experiment, run_id, name, value)
+
+    def record_settings(self, model, version, experiment, run_id, settings):
+        print("LocalState : record_settings")
+
+    def record_output(self, model, version, experiment, run_id, filename, file):
+        print("LocalState : record_output")
+
+    def record_session_start(self, model, version, experiment, run_id):
+        print("LocalState : record_session_start")
+
+    def record_session_end(self, model, version, experiment, run_id, status):
+        print("LocalState : record_session_end")
+
+
+class RemoteState(ObservatoryState):
+    #Handels metrics for the remote endpoint
 
     def _verify_response(self, response, expected_status, expected_type='application/json'):
         """
@@ -260,126 +240,170 @@ class TrackingSession:
             raise RuntimeError(f'Failed to execute operation. ' +
                                'Received invalid response type: {actual_type}')
 
-    def record_metric(self, name, value):
+    def record_metric(self, model, version, experiment, run_id, name, value):
         """
-        Records a metric value on the server
+        Records a metric during a run.
 
-        Parameters
-        ----------
-        name : string
-            The name of the metric to record
-        value : float
-            The value of the metric to records
+        This method sends a HTTP request with the right payload to the observatory tracking endpoint.
+        The result is a 201 when the server succesfully recorded the session completion. Otherwise
+        the server will return a 500 response.
+
+        Parameters:
+        -----------
+        model : str
+            The name of the model
+        version : int
+            The version of the model
+        experiment : str
+            The name of the experiment
+        run_id : str
+            The identifier for the run
+        metric_name : str
+            The name of the metric
+        metric_value : str
+            The value of the metric
+
+        Returns:
+        --------
+        requests.Response
+            The response from the server
         """
+        handler_url = f'{settings.server_url}/api/models/{model}/versions/{version}/experiments/{experiment}/runs/{run_id}/metrics'
+        self._verify_response(requests.post(handler_url, json={'name': name, 'value': value}), 201)
 
-        # Typechecking in python is a no-go under normal circumstances.
-        # But here we're using it, because the server expects a string and float.
-
-        if name is None or type(name) != str or name.strip() == '':
-            raise AssertionError('Please provide a valid name for the metric.')
-
-        if not re.match(LABEL_PATTERN, name):
-            raise AssertionError(
-                'Please provide a valid name for the metric.' +
-                'it can contain lower-case alpha-numeric characters and dashes only.')
-
-        if value is None or (type(value) != float and type(value) != int):
-            raise AssertionError(
-                'Please provide a valid value for the metric.')
-
-        response = self.tracking_client.record_metric(
-            self.name, self.version, self.experiment, self.run_id, name, value)
-
-        self._verify_response(response, 201)
-
-    def record_settings(self, **settings):
+    def record_settings(self, model, version, experiment, run_id, name, value):
+        print("RemoteState : record_settings")
         """
-        Records settings used for the run
+        Records the settings of an experiment run.
 
-        Parameters
-        ----------
-        settings : object
-            A dictionary containing all settings used for the run.
-            This can be passed in as `key=value` pairs.
+        This method sends a HTTP request with the right payload to the observatory tracking endpoint.
+        The result is a 201 when the server succesfully recorded the session completion. Otherwise
+        the server will return a 500 response.
+
+        Parameters:
+        -----------
+        model : str
+            The name of the model
+        version : int
+            The version of the model
+        experiment : str
+            The name of the experiment
+        run_id : str
+            The identifier for the run
+        settings : dict
+            The dictionary with run settings
+
+        Returns:
+        --------
+        requests.Response
+            The response from the server
         """
-        if settings is None:
-            warnings.warn(
-                'Trying to record empty settings. ' + 
-                'To prevent risk of settings loss on the server, ' +
-                'the empty settings collection is discarded.', 
-                RuntimeWarning)
+        handler_url = f'{settings.server_url}/api/models/{model}/versions/{version}/experiments/{experiment}/runs/{run_id}/settings'
+        self._verify_response(requests.post(handler_url, json=settings), 201)
+        
 
-        response = self.tracking_client.record_settings(
-            self.name, self.version, self.experiment,
-            self.run_id, dict(settings))
 
-        self._verify_response(response, 201)
-
-    def record_output(self, input_file, filename):
+    def record_output(self, model, version, experiment, run_id, filename, value):
+        print("RemoteState : record_output")
         """
-        Records an output for the current run.
+        Records an output of an experiment run
 
-        Parameters
-        ----------
-        input_file : object
-            Filename or handle to input file
-        filename : string
-            Name of the file as it should be stored on the server
+        This method sends a HTTP request with the right payload to the observatory tracking endpoint.
+        The result is a 201 when the server succesfully recorded the session completion. Otherwise
+        the server will return a 500 response.
+
+        Parameters:
+        -----------
+        model : str
+            The name of the model
+        version : int
+            The version of the model
+        experiment : str
+            The name of the experiment
+        run_id : str
+            The identifier for the run
+        filename : str
+            The filename of the output
+        file : object
+            The file handle to use for reading the output data
+
+        Returns:
+        --------
+        requests.Response
+            The response from the servers
         """
+        handler_url = f'{settings.server_url}/api/models/{model}/versions/{version}/experiments/{experiment}/runs/{run_id}/outputs/{filename}'
 
-        if filename is None or filename.strip() == '':
-            raise AssertionError(
-                'Please provide a valid filename to store the output on the server.')
+        file_collection = {
+            'file': (filename, file, 'application/octet-stream')
+        }
 
-        if input_file is None or input_file.strip() == '':
-            raise AssertionError(
-                'Please provide a valid filename for input_file.')
+        self._verify_response(requests.put(handler_url, files=file_collection), 201)
 
-        absolute_file_path = path.abspath(input_file)
+    def record_session_start(self, model, version, experiment, run_id):
+        """
+        Records the start of a session.
 
-        if not path.exists(absolute_file_path):
-            raise AssertionError(
-                f'Could not find source file {absolute_file_path} ' +
-                'to upload as output of this run. Please make ' +
-                'sure that the file exists on disk.')
+        This method sends a HTTP request with the right payload to the observatory tracking endpoint.
+        The result is a 201 when the server succesfully recorded the session completion. Otherwise
+        the server will return a 500 response.
 
-        response = self.tracking_client.record_output(
-            self.name, self.version, self.experiment,
-            self.run_id, filename, open(absolute_file_path))
+        Parameters:
+        -----------
+        model : str
+            The name of the model
+        version : int
+            The version of the model
+        experiment : str
+            The name of the experiment
+        run_id : str
+            The identifier for the run
 
-        self._verify_response(response, 201)
+        Returns:
+        --------
+        requests.Response
+            The response from the server
+        """
+        handler_url = f'{settings.server_url}/api/models/{model}/versions/{version}/experiments/{experiment}/runs'
+        self._verify_response(requests.post(handler_url, json={'run_id': run_id}), 201)
 
-    def __enter__(self):
-        response = self.tracking_client.record_session_start(
-            self.name, self.version, self.experiment, self.run_id)
+    def record_session_end(self, model, version, experiment, run_id, status):
+        """
+        Records the end of a session.
 
-        self._verify_response(response, 201)
+        This method sends a HTTP request with the right payload to the observatory tracking endpoint.
+        The result is a 201 when the server succesfully recorded the session completion. Otherwise
+        the server will return a 500 response.
 
-        return self
+        Parameters:
+        -----------
+        model : str
+            The name of the model
+        version : int
+            The version of the model
+        experiment : str
+            The name of the experiment
+        run_id : str
+            The identifier for the run
+        status : str
+            The status of the run
 
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        if exc_type is None:
-            session_status = 'COMPLETED'
-        else:
-            session_status = 'FAILED'
+        Returns:
+        --------
+        requests.Response
+            The response from the server
+        """
+        handler_url = f'{settings.server_url}/api/models/{model}/versions/{version}/experiments/{experiment}/runs/{run_id}'
+        self._verify_response(requests.put(handler_url, json={'status': status}), 201)
 
-        response = self.tracking_client.record_session_end(
-            self.name, self.version, self.experiment,
-            self.run_id, session_status)
-
-        self._verify_response(response, 201)
-
-        return exc_type is None
-
-
-def start_run(model, version, experiment='default'):
+def start_run(model, version, state=LocalState(), experiment='default'):
     """
     Starts a new run for a specific model version.
 
     You can use this function start start recording data related to a single experiment run.
     The use of experiments is optional, you can leave this parameter out. See the example below for basic usage.
 
-    >>> with observatory.start_run('my_model', 1, experiment='my_experiment') as run:
+    >>> with observatory.start_run('my_model', 1, remote, experiment='my_experiment') as run:
     >>>     # Record metrics, outputs and settings
 
     A new run is created on the server at the start of the scope.
@@ -388,7 +412,7 @@ def start_run(model, version, experiment='default'):
     Note that when an exception occurs within the scope of a run,
     the run is automatically marked as failed.
 
-    Each time you invoke `start_run` a new run is created with a newly generated UUID.
+    Each time you invoke `start_run()` a new run is created with a newly generated UUID.
     This ensures that no runs overlap eachother.
 
     **Please note** it is not possible to resume a run that you started earlier.
@@ -399,6 +423,8 @@ def start_run(model, version, experiment='default'):
         The name of the model
     version : int
         The version number of the model
+    state : object, optinal
+        The state you're working in
     experiment : string, optional
         The experiment you're working on
     """
@@ -420,8 +446,14 @@ def start_run(model, version, experiment='default'):
 
     if version <= 0:
         raise AssertionError('version must be greater than zero')
+    
+    if state != isinstance(state, (LocalState, RemoteState)):
+        raise AssertionError('State must be either a LocalState() or RemoteState()')
 
     run_id = str(uuid4())
-    tracking_client = TrackingClient(settings.server_url)
 
-    return TrackingSession(model, version, experiment, run_id, tracking_client)
+    trackingSession = TrackingSession(model, version, experiment, run_id)
+
+    trackingSession.change(state)
+    
+    return trackingSession
